requirements.txt
pytest==8.4.1
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.4
python-dotenv==1.1.1